Add Spectrum Next support
New binaries are created for the Next.  These are all dotn files
rather than .tap files.  All the new Next opcodes are supported
in addition to one pseudo op, nbrk, which is aliases to nextreg
2, 8 which starts the next debugger.

Next .x file and 48Kb .x files are sadly not compatible.  The
Next opcodes are not available in the 48k builds.

No new functionality is added for now to the Next dotn files,
even though more memory is available.

salink and specasm run at 28Mhz but restore the CPU to the user
selected speed on exit.  They also force the Next into the 48k
screen mode, which is restored naturally by the Next when the
programs exit.

+++ tests/test_sub_too_big/test.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 
 set -e
+
+if [ -z "${SPECASM_TARGET_NEXT_OPCODES}" ]; then
+    exit 0
+fi
+
 rm subtraction 2>/dev/null 1>&2 || true
 
 if ../../salink 2>/dev/null 1>&2 ; then
